Splitted TTe and eTT modules, all constants collected constants.py file, added 2 fake tuples inserter to simulate stt and erasmus_nlp groups for debugging purposes
