--- SmartApp.ENLP/coords.py
@@ -1,11 +1,0 @@
-# emotion : (valence, arousal)
-EMOTIONS_COORD = {
-    'Excited' : (0.5,0.8),
-    'Angry' : (-0.8,0.5),
-    'Bored' : (-0.5,-0.8),
-    'Happy' : (0.8,0.5),
-    'Fear' : (-0.6,0.8),
-    'Sarcasm' : (0.,0.3),
-    'Sad' : (-0.9,-0.6),
-    'Neutral' : (0.,0.)
-}
